requirements.txt
blessed==1.20.0
PyYAML==6.0.1
